+++ karbide-core/src/commonMain/kotlin/dev/karmakrafts/karbide/BitSinkExtensions.kt
@@ -29,25 +29,11 @@ fun BitSink.writeByte(value: Byte) = writeBits(Byte.SIZE_BITS, value.toULong())
 fun BitSink.writeShort(value: Short) = writeBits(Short.SIZE_BITS, value.toULong())
 
 /**
- * Write a [Short] (16 bits) to the [BitSink] in little endian byte order.
- *
- * @param value The short value to write.
- */
-fun BitSink.writeShortLe(value: Short) = writeBits(Short.SIZE_BITS, value.reverseBytes().toULong())
-
-/**
  * Write an [Int] (32 bits) to the [BitSink].
  *
  * @param value The integer value to write.
  */
 fun BitSink.writeInt(value: Int) = writeBits(Int.SIZE_BITS, value.toULong())
-
-/**
- * Write an [Int] (32 bits) to the [BitSink] in little endian byte order.
- *
- * @param value The integer value to write.
- */
-fun BitSink.writeIntLe(value: Int) = writeBits(Int.SIZE_BITS, value.reverseBytes().toULong())
 
 /**
  * Write a [Long] (64 bits) to the [BitSink].
@@ -57,13 +43,6 @@ fun BitSink.writeIntLe(value: Int) = writeBits(Int.SIZE_BITS, value.reverseBytes
 fun BitSink.writeLong(value: Long) = writeBits(Long.SIZE_BITS, value.toULong())
 
 /**
- * Write a [Long] (64 bits) to the [BitSink] in little endian byte order.
- *
- * @param value The long value to write.
- */
-fun BitSink.writeLongLe(value: Long) = writeBits(Long.SIZE_BITS, value.reverseBytes().toULong())
-
-/**
  * Write a [Float] (32 bits) to the [BitSink].
  *
  * @param value The float value to write.
@@ -71,25 +50,11 @@ fun BitSink.writeLongLe(value: Long) = writeBits(Long.SIZE_BITS, value.reverseBy
 fun BitSink.writeFloat(value: Float) = writeBits(Float.SIZE_BITS, value.toRawBits().toULong())
 
 /**
- * Write a [Float] (32 bits) to the [BitSink] in little endian byte order.
- *
- * @param value The float value to write.
- */
-fun BitSink.writeFloatLe(value: Float) = writeBits(Float.SIZE_BITS, value.toRawBits().reverseBytes().toULong())
-
-/**
  * Write a [Double] (64 bits) to the [BitSink].
  *
  * @param value The double value to write.
  */
 fun BitSink.writeDouble(value: Double) = writeBits(Double.SIZE_BITS, value.toRawBits().toULong())
-
-/**
- * Write a [Double] (64 bits) to the [BitSink] in little endian byte order.
- *
- * @param value The double value to write.
- */
-fun BitSink.writeDoubleLe(value: Double) = writeBits(Double.SIZE_BITS, value.toRawBits().reverseBytes().toULong())
 
 /**
  * Write a [UByte] (8 bits) to the [BitSink].
@@ -106,13 +71,6 @@ fun BitSink.writeUByte(value: UByte) = writeBits(UByte.SIZE_BITS, value.toULong(
 fun BitSink.writeUShort(value: UShort) = writeBits(UShort.SIZE_BITS, value.toULong())
 
 /**
- * Write a [UShort] (16 bits) to the [BitSink] in little endian byte order.
- *
- * @param value The unsigned short value to write.
- */
-fun BitSink.writeUShortLe(value: UShort) = writeBits(UShort.SIZE_BITS, value.reverseBytes().toULong())
-
-/**
  * Write a [UInt] (32 bits) to the [BitSink].
  *
  * @param value The unsigned integer value to write.
@@ -120,25 +78,11 @@ fun BitSink.writeUShortLe(value: UShort) = writeBits(UShort.SIZE_BITS, value.rev
 fun BitSink.writeUInt(value: UInt) = writeBits(UInt.SIZE_BITS, value.toULong())
 
 /**
- * Write a [UInt] (32 bits) to the [BitSink] in little endian byte order.
- *
- * @param value The unsigned integer value to write.
- */
-fun BitSink.writeUIntLe(value: UInt) = writeBits(UInt.SIZE_BITS, value.reverseBytes().toULong())
-
-/**
  * Write a [ULong] (64 bits) to the [BitSink].
  *
  * @param value The unsigned long value to write.
  */
 fun BitSink.writeULong(value: ULong) = writeBits(ULong.SIZE_BITS, value)
-
-/**
- * Write a [ULong] (64 bits) to the [BitSink] in little endian byte order.
- *
- * @param value The unsigned long value to write.
- */
-fun BitSink.writeULongLe(value: ULong) = writeBits(ULong.SIZE_BITS, value.reverseBytes())
 
 /**
  * Write a [ByteArray] to the [BitSink].
@@ -150,6 +94,13 @@ fun BitSink.writeByteArray(value: ByteArray) {
         writeByte(byte)
     }
 }
+
+/**
+ * Write a [String] to the [BitSink].
+ *
+ * @param value The string value to write.
+ */
+fun BitSink.writeString(value: String) = writeByteArray(value.encodeToByteArray())
 
 /**
  * Write a single padding bit to the [BitSink].